requirements.txt
boto3
redis
river
ruptures
websocket-client
python-dotenv
fastapi
uvicorn
pandas
numpy
mangum
yfinance
stripe
requests
